requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.14.0
azure-functions-worker>=1.0.0

# Data processing
pandas>=1.5.0
openpyxl>=3.1.0

# AI/ML
openai>=1.3.0

# Code analysis and validation
ast-grep>=0.15.0
flake8>=6.0.0
black>=23.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
mypy>=1.5.0
pre-commit>=3.0.0

# Utilities
requests>=2.25.0
python-dotenv>=0.19.0
hashlib2>=1.0.0
pathlib>=1.0.0
